moderation-service/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic-settings>=2.3,<3.0
requests>=2.32,<3.0
Pillow>=10.4,<12.0
transformers>=4.45,<5.0
torch>=2.3,<3.0
